docs: completa comentarios Javadoc en espanol para las 43 entidades con trazabilidad 1:1 a base_datos.sql y actualiza estado_sistema_v.0.3.md

+++ src/main/java/com/app/idoneos/controller/DescuentoController.java
@@ -56,7 +56,7 @@ public class DescuentoController {
 
         Descuento d = new Descuento();
         d.setNombre(nombre);
-        d.setPorcentaje(porcentaje);
+        d.setPorcentaje(porcentaje.floatValue());
         d.setVigenciaDesde(vigenciaDesde);
         d.setVigenciaHasta(vigenciaHasta);
         d.setCantidadLimite(cantidadLimite);
@@ -96,7 +96,7 @@ public class DescuentoController {
 
         Descuento d = dOpt.get();
         d.setNombre(nombre);
-        d.setPorcentaje(porcentaje);
+        d.setPorcentaje(porcentaje.floatValue());
         d.setVigenciaDesde(vigenciaDesde);
         d.setVigenciaHasta(vigenciaHasta);
         d.setCantidadLimite(cantidadLimite);
